Move back pattern groups resolution to GitSituation
